requirement.txt
streamlit
requests
beautifulsoup4
groq
python-dotenv